refactor: move SetIdentifierArgumentFn to config/externalname.go
Move the cold-start fix from per-package config files
(config/cluster/ecs/config.go, config/namespaced/ecs/config.go) to
config/externalname.go as a custom ExternalName config function.

This follows repository convention: externalname.go is the single
source of truth for external name configurations, deduplicating
across cluster and namespaced variants.
