+++ config/externalname_ecs_test.go
@@ -2,33 +2,34 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package ecs
+package config
 
 import (
 	"testing"
 )
 
-func TestTaskDefinitionSetIdentifierArgumentFn(t *testing.T) {
-	setIDFn := func(base map[string]any, externalName string) {
-		if arn, _ := base["arn"].(string); arn == "" && externalName != "" {
-			base["arn"] = externalName
-		}
-	}
+func TestEcsTaskDefinitionSetIdentifierArgumentFn(t *testing.T) {
+	e := ecsTaskDefinition()
 
 	cases := map[string]struct {
 		base         map[string]any
 		externalName string
 		wantArn      string
 	}{
-		"ColdStartObserveWithFullARN": {
+		"ColdStartWithFullARN": {
 			base:         map[string]any{},
 			externalName: "arn:aws:ecs:us-east-1:123456789012:task-definition/my-service:7",
 			wantArn:      "arn:aws:ecs:us-east-1:123456789012:task-definition/my-service:7",
 		},
-		"ColdStartObserveWithFamilyRevision": {
+		"ColdStartWithFamilyRevision": {
 			base:         map[string]any{},
 			externalName: "my-service:7",
 			wantArn:      "my-service:7",
+		},
+		"ColdStartWithFamilyOnly": {
+			base:         map[string]any{},
+			externalName: "my-service",
+			wantArn:      "my-service",
 		},
 		"WarmStartArnAlreadyPopulated": {
 			base:         map[string]any{"arn": "arn:aws:ecs:us-east-1:123456789012:task-definition/my-service:7"},
@@ -44,7 +45,7 @@ func TestTaskDefinitionSetIdentifierArgumentFn(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			setIDFn(tc.base, tc.externalName)
+			e.SetIdentifierArgumentFn(tc.base, tc.externalName)
 			got, _ := tc.base["arn"].(string)
 			if got != tc.wantArn {
 				t.Errorf("base[\"arn\"] = %q, want %q", got, tc.wantArn)
